requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==6.0.3
django-fernet-encrypted-fields==0.4.0
django-picklefield==3.4.0
django-q2==1.9.0
idna==3.11
psycopg2==2.9.11
pycparser==3.0
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
urllib3==2.6.3
